Add new function Compound.Bar (#4)
Adds a new function that creates bar chart, using the similar pattern to the Line or Area charts. The min and max for x axis is based on the measures for all selected values on the x axis, rather than specific to the filtered dimension, so that comparisons across the charts can be more easily made.